requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
numpy==1.24.3
opencv-python==4.8.1.78